backend/requirements.txt
asgiref==3.10.0
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
sqlparse==0.4.4
whitenoise==6.6.0
gunicorn==21.2.0
python-dotenv==1.0.0
sentry-sdk==1.40.6
Pillow==10.1.0
psycopg2-binary==2.9.9
argon2-cffi==23.1.0
django-otp==1.1.3
django-ratelimit==4.1.0
defusedxml>=0.7.1
